pyproject.toml
[project]
name = "strands-playground"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "anthropic[vertex]>=0.75.0",
    "boto3>=1.42.18",
    "google-cloud-aiplatform>=1.132.0",
    "opensearch-py>=3.0.0",
    "pandas>=2.3.3",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "strands-agents[litellm]>=1.20.0",
    "strands-agents-tools[tavily]>=0.2.18",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[project.optional-dependencies]
rerank = [
    "rerankers[flashrank]",
]

[tool.ruff]
line-length = 120
target-version = "py311"
src = ["src", "scripts", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
ignore = ["E501"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "ruff>=0.14.10",
]

[tool.pytest.ini_options]
filterwarnings = [
    "ignore::UserWarning",
    "ignore::urllib3.exceptions.InsecureRequestWarning",
]
testpaths = ["tests"]

[project.scripts]
test-opensearch = "scripts.test_opensearch:main"
rag-chat = "src.cli.main:main"
